refactor: ♻️ move preview_repository to repositories layer (#264)
