+++ backend/src/services/mod.rs
@@ -7,7 +7,8 @@ pub mod github_pr_service;
 pub mod github_sync_service;
 pub mod invitation_service;
 pub mod member_service;
-pub mod preview_repository;
+/// Re-export from repositories layer for backward compatibility.
+pub use crate::repositories::preview_repository;
 pub mod preview_service;
 pub mod project_message_service;
 pub mod project_service;
